Don't use module attributes for configuration
Module attributes are statically compiled, which for us happens when
running the Docker build. Therefore the correct configuration won't be
present, so the result of `@config` is an empty map. This causes the
code to fail.

The configuration is present at runtime though, so the fix is to use a
function instead of a compiled module attribute.